Rename CheckSetup files to check_setup
