+++ vpr/src/base/check_setup.h
